task: tidy project root: webpack
This change attempts to tidy the project root by moving webpack configuration
into it's own subdirectory.  The metadata has been updated to match this information.

In addition, the configuration file itself is simplified for future maintainers
in some ways. Now the configuration to a projectRoot which is calculated
exactly once, instead of introspecting from __dirname each time.